@@ -93,33 +97,34 @@
     "pprof": "^3.2.0",
     "prom-client": "^14.2.0",
     "qs": "^6.11.1",
-    "rudder-transformer-cdk": "^1.4.5",
+    "rudder-transformer-cdk": "^1.4.9",
     "rudder-workflow-engine": "^0.4.6",
     "set-value": "^4.1.0",
     "sha256": "^0.2.0",
     "stacktrace-parser": "^0.1.10",
     "statsd-client": "^0.4.7",
     "truncate-utf8-bytes": "^1.0.2",
-    "ua-parser-js": "^1.0.34",
+    "ua-parser-js": "^1.0.35",
     "unset-value": "^2.0.1",
     "uuid": "^9.0.0",
     "valid-url": "^1.0.9"
   },
   "devDependencies": {
-    "@commitlint/config-conventional": "^17.4.4",
-    "@digitalroute/cz-conventional-changelog-for-jira": "^8.0.0",
-    "@types/jest": "^29.5.0",
+    "@commitlint/config-conventional": "^17.6.3",
+    "@digitalroute/cz-conventional-changelog-for-jira": "^8.0.1",
+    "@types/jest": "^29.5.1",
     "@types/koa": "^2.13.6",
     "@types/koa-bodyparser": "^4.3.10",
+    "@types/node": "^20.2.5",
     "@typescript-eslint/eslint-plugin": "^5.59.2",
     "@typescript-eslint/parser": "^5.59.2",
     "acorn": "^8.8.2",
     "benchmark-suite": "^0.1.8",
-    "commander": "^10.0.0",
+    "commander": "^10.0.1",
     "commitizen": "^4.3.0",
-    "commitlint": "^17.4.4",
+    "commitlint": "^17.6.3",
     "conventional-github-releaser": "^3.1.5",
-    "eslint": "^8.36.0",
+    "eslint": "^8.40.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
