@@ -328,13 +328,8 @@ async function userTransformHandler(
   return events;
 }
 
-async function setupUserTransformHandler(
-  trRevCode = {},
-  libraryVersionIDs,
-) {
-  const resp = await UserTransformHandlerFactory(trRevCode).setUserTransform(
-    libraryVersionIDs
-  );
+async function setupUserTransformHandler(libraryVersionIDs, trRevCode = {}) {
+  const resp = await UserTransformHandlerFactory(trRevCode).setUserTransform(libraryVersionIDs);
   return resp;
 }
 
