@@ -6,9 +6,8 @@ const stats = require('../stats');
 
 const timeoutPromise = () =>
   new Promise((_, reject) => {
-    setTimeout(() => reject(new Error('Timeout while connecting to redis')), 1000);
+    setTimeout(() => reject(new Error('Timeout while connecting to redis')), 1200);
   });
-
 const RedisDB = {
   init() {
     if (process.env.USE_REDIS_DB && process.env.USE_REDIS_DB !== 'false') {
@@ -34,6 +33,7 @@ const RedisDB = {
           log.error(`Redis is down at ${this.host}:${this.port}`);
           return false; // stop retrying
         },
+        tls: {}
       });
       this.client.on('ready', () => {
         log.info(`Connected to redis at ${this.host}:${this.port}`);
